Add toString() implementation to Asset and move Claimant to org.stellar.sdk package
